# webapp/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: VooZuu App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/forgotPassword:
    # binds a127 app logic to a route
    x-swagger-router-controller: auth_ctrl
    post:
      tags: ['Auth Module']
      description: User Forgot Password
      summary: User Forgot Password
      # used as the method name of the controller
      operationId: forgotPassword
      parameters:
        - name: body
          in: body
          description: Forgot Password
          schema:
           $ref: '#/definitions/forgotPassword'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/resetPassword:
    # binds a127 app logic to a route
    x-swagger-router-controller: auth_ctrl
    post:
      tags: ['Auth Module']
      description: Admin Reset Password
      summary: Admin Reset Password
      # used as the method name of the controller
      operationId: resetPassword
      parameters:
        - name: body
          in: body
          description: Reset Password
          schema:
           $ref: '#/definitions/resetPassword'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/adminLoggedin:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminAuth_ctrl
    get:
      tags: ['Admin Auth Module']
      description: To check admin loggedin or not
      summary: To check admin loggedin or not
      # used as the method name of the controller
      operationId: adminLoggedin   
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/adminLogin:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminAuth_ctrl
    post:
      tags: ['Admin Auth Module']
      description: Admin login
      summary:  Admin login
      # used as the method name of the controller
      operationId: adminLogin
      parameters:
        - name: body
          in: body
          description:  Admin login
          schema:
           $ref: '#/definitions/adminLogin'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/adminForgotPassword:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminAuth_ctrl
    post:
      tags: ['Admin Auth Module']
      description: Admin Forgot Password
      summary: Admin Forgot Password
      # used as the method name of the controller
      operationId: adminForgotPassword
      parameters:
        - name: body
          in: body
          description: Forgot Password
          schema:
           $ref: '#/definitions/forgotPassword'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/adminResetPassword:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminAuth_ctrl
    post:
      tags: ['Admin Auth Module']
      description: Admin Reset Password
      summary: Admin Reset Password
      # used as the method name of the controller
      operationId: adminResetPassword
      parameters:
        - name: body
          in: body
          description: Reset Password
          schema:
           $ref: '#/definitions/resetPassword'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getAdminById/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminAuth_ctrl
    get:
      tags: ['Admin Auth Module']
      description: Get Admin By Id 
      summary:  Get Admin By Id 
      # used as the method name of the controller
      operationId: getAdminById      
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          type: string        
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                         
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/adminProfileUpdate:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminAuth_ctrl
    post:
      tags: ['Admin Auth Module']
      description: Update Admin Profile Info
      summary: Update Admin Profile Info
      # used as the method name of the controller
      operationId: adminProfileUpdate
      parameters:
        - name: body
          in: body
          description: Update User Profile Info
          schema:
           $ref: '#/definitions/adminProfileUpdate'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/adminChangePassword:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminAuth_ctrl
    post:
      tags: ['Admin Auth Module']
      description: Change User Password
      summary: Change User Password
      # used as the method name of the controller
      operationId: adminChangePassword
      parameters:
        - name: body
          in: body
          description: Change User Password
          schema:
           $ref: '#/definitions/adminChangePassword'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/adminUpdateImage:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminAuth_ctrl
    post:
      tags: ['Admin Auth Module']
      description: Update Admin Profile Pic
      summary: Update Admin Profile Pic
      # used as the method name of the controller
      operationId: adminUpdateImage
      consumes:
        - multipart/form-data
        - application/x-www-form-urlencoded
        - binary
      parameters:
        - in: formData
          name: id
          description: User id
          type: string
          required: true                
        - in: formData
          name: file
          description: Update Admin Profile Pic
          type: file
          required: true                      
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  forgotPassword:
    type: object
    properties:
      email:
        type: string
  resetPassword:
    type: object
    properties:
      resetKey:
        type: string
      password:
        type: string
...
  adminLogin:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
  adminProfileUpdate:
    type: object
    properties:
      userId:
        type: string
      firstname:
        type: string
      lastname:
        type: string
  adminChangePassword:
    type: object
    properties:
      userId:
        type: string
      oldPassword:
        type: string
      newPassword:
        type: string  
...